Update script.js
This new script uses a flag to fix an issue I was having with the timer that was allowing the user to input multiple different times at the same time. This would break the timer as it would try and keep track of multiple different times at the same time. The flag variable called isRunning checks to see if the clock is currently running or not so that the user can only input one time at a time.